reformatting code to make it easier to understand the structure
